requirements.txt
pydantic>=2.0.0
rich>=13.0.0
pytest>=7.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
openai>=1.0.0
chromadb>=0.4.0
scipy>=1.10.0
tenacity>=8.0.0
